Update .gitignore
I made change to .gitigne as instructed on item 5. of Setting up the development environment for your project:
"The .gitignore file contains a list of file types that should be
ignored when the project is committed, or pushed to the
repository. Select the .gitignore file list below to modify it. This
file is necessary to avoid conflicts that might occur when team
members have different system configurations and operating
systems."
Jean Oliveira